+++ examples/chapi-example/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @learncard/chapi-example
 
+## 1.1.16
+
+### Patch Changes
+
+-   Updated dependencies [[`bf4f00306f64e701f3c9acee4c5f7438d3f3b6ee`](https://github.com/learningeconomy/LearnCard/commit/bf4f00306f64e701f3c9acee4c5f7438d3f3b6ee)]:
+    -   @learncard/types@5.12.3
+    -   @learncard/helpers@1.2.9
+    -   @learncard/init@2.3.7
+    -   @learncard/chapi-plugin@1.1.9
+    -   @learncard/react@2.9.16
+
+## 1.1.15
+
+### Patch Changes
+
+-   Updated dependencies [[`495f2939cb6e4271cab0a88abea5105fb7e4f9b6`](https://github.com/learningeconomy/LearnCard/commit/495f2939cb6e4271cab0a88abea5105fb7e4f9b6), [`08a1c8a07501c2f426d16239c4b0551e518a7ed5`](https://github.com/learningeconomy/LearnCard/commit/08a1c8a07501c2f426d16239c4b0551e518a7ed5)]:
+    -   @learncard/types@5.12.2
+    -   @learncard/helpers@1.2.8
+    -   @learncard/init@2.3.6
+    -   @learncard/chapi-plugin@1.1.8
+    -   @learncard/react@2.9.15
+
 ## 1.1.14
 
 ### Patch Changes