+++ packages/plugins/open-badge-v2/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @learncard/open-badge-v2-plugin
 
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies [[`bf4f00306f64e701f3c9acee4c5f7438d3f3b6ee`](https://github.com/learningeconomy/LearnCard/commit/bf4f00306f64e701f3c9acee4c5f7438d3f3b6ee)]:
+    -   @learncard/types@5.12.3
+    -   @learncard/core@9.4.9
+    -   @learncard/vc-plugin@1.4.5
+    -   @learncard/crypto-plugin@1.1.9
+
+## 1.1.9
+
+### Patch Changes
+
+-   Updated dependencies [[`495f2939cb6e4271cab0a88abea5105fb7e4f9b6`](https://github.com/learningeconomy/LearnCard/commit/495f2939cb6e4271cab0a88abea5105fb7e4f9b6), [`08a1c8a07501c2f426d16239c4b0551e518a7ed5`](https://github.com/learningeconomy/LearnCard/commit/08a1c8a07501c2f426d16239c4b0551e518a7ed5)]:
+    -   @learncard/types@5.12.2
+    -   @learncard/core@9.4.8
+    -   @learncard/vc-plugin@1.4.4
+    -   @learncard/crypto-plugin@1.1.8
+
 ## 1.1.8
 
 ### Patch Changes