+++ packages/plugins/simple-signing-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # lca-api-plugin
 
+## 1.1.11
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-client@1.1.11
+
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-client@1.1.10
+
 ## 1.1.9
 
 ### Patch Changes