+++ packages/react-learn-card/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @welibraryos/react-learn-card
 
+## 2.9.16
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.3.7
+
+## 2.9.15
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.3.6
+
 ## 2.9.14
 
 ### Patch Changes